requirements.txt
Flask==2.1.0
Werkzeug==2.0.2
python-dotenv==1.0.1
webargs==8.4.0
pytest==8.0.2
